continued bug fix for seg fault.  To prevent seg fault, make sure you start off with enable Tx disabled (eg by deleting the sdr-shell.conf file in ~/.config/freesoftware).  Once sdr-shell is started, you can click on CFG > TX > enable Tx button to enable Tx.  However, before you exit sdr-shell, you have to disable the "enable Tx" button.
This is because there is still an elusive bug during the startup
of sdr-shell.  With enable Tx enabled during startup, sdr-shell goes
into seg fault.  I have traced it to a problem with the DttSPTXCmd class.
For example, main_widget.cpp was still calling DttSPTXCmd: on(), which
seems to be incompatible with the new DttSPCmd class.  I have therefore
commented out calls to DttSPTXCmd: on(), with some reduction in likelihood
of seg faults.